n3_server/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# OpenTelemetry instrumentation
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-exporter-otlp>=1.21.0
openinference-instrumentation-openai>=0.1.0

# RLHF/PPO for adaptive policies
trl>=0.7.0
transformers>=4.35.0
torch>=2.1.0
peft>=0.7.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.8.2
nanoid>=2.0.0
httpx>=0.25.0

# Existing N3 dependencies (from requirements.txt)
langchain>=0.1.0
langchain-community>=0.0.10
openai>=1.3.0
anthropic>=0.8.0
google-cloud-aiplatform>=1.38.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
